+++ src/compute-engine/compile.ts
@@ -25,21 +25,23 @@ import {
 
 export type CompiledType = boolean | number | string | object;
 
-type JSSource = string;
+export type JSSource = string;
 
 export type CompiledOperators = Record<
   MathJsonIdentifier,
   [op: string, prec: number]
 >;
 
+export type CompiledFunction =
+  | string
+  | ((
+      args: ReadonlyArray<BoxedExpression>,
+      compile: (expr: BoxedExpression) => JSSource,
+      target: CompileTarget
+    ) => JSSource);
+
 export type CompiledFunctions = {
-  [id: MathJsonIdentifier]:
-    | string
-    | ((
-        args: ReadonlyArray<BoxedExpression>,
-        compile: (expr: BoxedExpression) => JSSource,
-        target: CompileTarget
-      ) => JSSource);
+  [id: MathJsonIdentifier]: CompiledFunction;
 };
 
 const NATIVE_JS_OPERATORS: CompiledOperators = {
@@ -360,13 +362,12 @@ const NATIVE_JS_FUNCTIONS: CompiledFunctions = {
 
 export type CompileTarget = {
   operators?: (op: MathJsonIdentifier) => [op: string, prec: number];
-  functions?: (
-    id: MathJsonIdentifier
-  ) => string | ((...args: CompiledType[]) => string);
+  functions?: (id: MathJsonIdentifier) => CompiledFunction | undefined;
   var: (id: MathJsonIdentifier) => string | undefined;
   string: (str: string) => string;
   number: (n: number) => string;
   ws: (s?: string) => string; // White space
+  preamble: string;
   indent: number;
   // @todo: add context or return compile as an array of statements
   // and let the caller decide how to wrap it in an IIFE.
@@ -400,8 +401,8 @@ export class ComputeEngineFunction extends Function {
     quartiles,
     interquartileRange,
   };
-  constructor(body: string) {
-    super('_SYS', '_', `return ${body}`);
+  constructor(body: string, preamble = '') {
+    super('_SYS', '_', `${preamble};return ${body}`);
     return new Proxy(this, {
       apply: (target, thisArg, argumentsList) =>
         super.apply(thisArg, [this.sys, ...argumentsList]),
@@ -418,24 +419,60 @@ export class ComputeEngineFunction extends Function {
 export function compileToTarget(
   expr: BoxedExpression,
   target: CompileTarget
-): ((_: Record<string, CompiledType>) => CompiledType) | undefined {
+): (_?: Record<string, CompiledType>) => CompiledType {
   const js = compile(expr, target);
-  // try {
-  return new ComputeEngineFunction(js) as unknown as () => CompiledType;
-  // } catch (e) {
-  //   console.error(`${e}\n${expr.latex}\n${js}`);
-  // }
-  // return undefined;
+  return new ComputeEngineFunction(
+    js,
+    target.preamble
+  ) as unknown as () => CompiledType;
 }
 
 export function compileToJavascript(
-  expr: BoxedExpression
-): ((_: Record<string, CompiledType>) => CompiledType) | undefined {
+  expr: BoxedExpression,
+  functions?: Record<MathJsonIdentifier, JSSource | Function>,
+  vars?: Record<MathJsonIdentifier, JSSource>,
+  imports: unknown[] = [],
+  preamble?: string
+): (_?: Record<string, CompiledType>) => CompiledType {
   const unknowns = expr.unknowns;
+
+  // For any import, turn it into a string
+  let preambleImports = imports
+    .map((x) => {
+      if (typeof x === 'function') return x.toString();
+      throw new Error(`Unsupported import \`${x}\``);
+    })
+    .join('\n');
+
+  // @ts-ignore
+  const namedFunctions: { [k: string]: string } = functions
+    ? Object.fromEntries(
+        Object.entries(functions).filter((k, v) => typeof v !== 'string')
+      )
+    : {};
+
+  if (functions)
+    for (const [k, v] of Object.entries(functions)) {
+      if (typeof v === 'function') {
+        // If a named function, turn it into a string
+        // otherwise, declare it as a constant
+
+        if (isTrulyNamed(v)) {
+          preambleImports += `${v.toString()};\n`;
+          namedFunctions[k] = v.name;
+        } else {
+          preambleImports += `const ${k} = ${v.toString()};\n`;
+          namedFunctions[k] = k;
+        }
+      }
+    }
+
   return compileToTarget(expr, {
     operators: (op) => NATIVE_JS_OPERATORS[op],
-    functions: (id) => NATIVE_JS_FUNCTIONS[id],
+    functions: (id) =>
+      namedFunctions?.[id] ? namedFunctions[id] : NATIVE_JS_FUNCTIONS[id],
     var: (id) => {
+      if (vars && id in vars) return JSON.stringify(vars[id]);
       const result = {
         Pi: 'Math.PI',
         ExponentialE: 'Math.E',
@@ -455,6 +492,7 @@ export function compileToJavascript(
     number: (n) => n.toString(),
     indent: 0,
     ws: (s?: string) => s ?? '',
+    preamble: (preamble ?? '') + preambleImports,
   });
 }
 
@@ -569,7 +607,7 @@ function compileExpr(
   }
 
   const fn = target.functions?.(h);
-  if (!fn) throw new Error(`Unknown function ${h}`);
+  if (!fn) throw new Error(`Unknown function \`${h}\``);
   if (typeof fn === 'function') {
     // Get function definition for h
     const def = engine.lookupFunction(h);
@@ -581,22 +619,15 @@ function compileExpr(
     ) {
       const v = tempVar();
       return `(${compile(args[0], target)}).map((${v}) => ${fn(
-        args[0].engine.box(v),
-        (expr) => compile(expr, target)
+        [args[0].engine.box(v)],
+        (expr) => compile(expr, target),
+        target
       )})`;
     }
     return fn(args, (expr) => compile(expr, target), target);
   }
 
   if (args === null) return `${fn}()`;
-
-  if (args.length === 1 && isFiniteIndexableCollection(args[0])) {
-    const v = tempVar();
-    return `(${compile(args[0], target)}).map((${v}) => ${fn}(${compile(
-      args[0].engine.box(v),
-      target
-    )}))`;
-  }
 
   return `${fn}(${args.map((x) => compile(x, target)).join(', ')})`;
 }
@@ -738,4 +769,14 @@ function compileIntegrate(args, _, target: CompileTarget): string {
   });
 
   return `_SYS.integrate((${index}) => (${f}), ${lower}, ${upper})`;
+}
+
+function isTrulyNamed(func: Function): boolean {
+  const source = func.toString();
+
+  // Check if it's an arrow function, which is always anonymous
+  if (source.includes('=>')) return false;
+
+  // Check if the function has a name in `.toString()`
+  return source.startsWith('function ') && source.includes(func.name);
 }